example/azure-loganalytics/python/requirements.txt
azure-loganalytics
msal
msrest
